Split namespace type into its own module
