+++ capnp.cabal
@@ -78,6 +78,8 @@ library
 
 executable capnpc-haskell
     main-is:              Main.hs
+    other-modules:
+      Namespace
     hs-source-dirs:       executables/capnpc-haskell
     build-depends:
         capnp
