+++ executables/capnpc-haskell/Main.hs
@@ -13,6 +13,7 @@ import Data.DList               (DList)
 import Data.Functor.Identity    (Identity(..))
 import Data.List                (intersperse)
 import Data.Monoid              (mconcat, (<>))
+import Namespace
 
 import qualified Data.ByteString                                                   as BS
 import qualified Data.CapNProto.BasicTypes                                         as BT
@@ -29,23 +30,10 @@ import qualified Schema.CapNProto.Reader.Schema.Node                            
 import qualified Schema.CapNProto.Reader.Schema.Node.NestedNode                    as NN
 
 type BS = BS.ByteString
-type NS = [BT.Text BS]
 
 type NodeMap = M.Map Node.Id (Schema.Node BS)
 
 type Generator a = CatchT (Writer (DList (NS, TH.DecsQ))) a
-
-moduleName :: NS -> BT.Text BS
-moduleName ns = mconcat $
-    "Schema.CapNProto.Reader." :
-    reverse (intersperse "." ns)
-
-moduleFile :: NS -> FilePath
-moduleFile ns = asStr $ mconcat ( "Schema/CapNproto/Reader/"
-                                : reverse (intersperse "/" ns)
-                                ) <> ".hs"
-  where
-    asStr (BT.Text path) = toString path
 
 buildNodeMap :: List.ListOf BS (Schema.Node BS) -> Generator NodeMap
 buildNodeMap = List.foldl addNode M.empty
